% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netflix_documentation.R
\docType{data}
\name{country_data}
\alias{country_data}
\title{Country data}
\format{
a data frame with Netflix movies in availble countries
\describe{
\item{show_id}{show of movie ID number}
\item{type}{TV Show or Movie}
\item{title}{Title of movie or tv show}
\item{director}{Director name}
\item{cast}{List of cast members}
\item{country}{Country where the movie was produced}
\item{date_added}{Date it was added to Netlix}
\item{release_year}{Release year of tv show or movie}
\item{rating}{Maturity rating of of tv show or movie}
\item{duration}{Integer duration of Movie or TV Show (in minutes if movie, in seasons if tv show)}
\item{listed_in}{List of applicable genres}
\item{description}{Description of movie}
}
}
\usage{
country_data
}
\description{
This dataset consists of tv shows and movies available on Netflix as of 2019. Unlike the netflix datset, the country column is split
so that the movie is repeated for how ever many countries it is availble to
}
\examples{
country_data
}
\keyword{datasets}
