% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netflix_documentation.R
\docType{data}
\name{country_listedin}
\alias{country_listedin}
\title{Country Listed-in data}
\format{
a data frame with Netflix movies in available countries
\describe{
\item{show_id}{show of movie ID number}
\item{type}{TV Show or Movie}
\item{title}{Title of movie or tv show}
\item{country}{Country where the movie was produced}
\item{release_year}{Release year of tv show or movie}
\item{rating}{Maturity rating of of tv show or movie}
\item{listed_in}{List of applicable genres}
\item{description}{Description of movie}
}
}
\usage{
country_listedin
}
\description{
This data set consists of TV shows and Movies available on Netflix as of 2019. Unlike the netflix dataset, the country column is split
so that the movie is repeated for how ever many countries it is available to. Unlike the country_data dataset, the listed_in column
is also split so that each row has a unique movie + country + listed in. The movie is repeated if it has more than one 'listed_in' and
if it has more than one country it us available to
}
\examples{
country_listedin
}
\keyword{datasets}
