% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netflix.R
\name{country_titles}
\alias{country_titles}
\title{Function for getting movie/tv list from a given country on production}
\usage{
country_titles(country)
}
\arguments{
\item{country}{Country of production}
}
\value{
None
}
\description{
Function for getting movie/tv list from a given country on production
}
\examples{
country_titles("United States")
}
