% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netflix.R
\name{country_year_titles}
\alias{country_year_titles}
\title{Function for getting movie/tv list from a given country on production}
\usage{
country_year_titles(country, year)
}
\arguments{
\item{country}{Country of production}

\item{year}{Year of release}
}
\value{
None
}
\description{
Function for getting movie/tv list from a given country on production
}
\examples{
country_year_titles("United States", 2020)
}
