% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netflix_documentation.R
\docType{data}
\name{netflix}
\alias{netflix}
\title{Netflix data}
\format{
A data frame with Movies and TV Shows available on Netflix
\describe{
\item{show_id}{show of movie ID number}
\item{type}{TV Show or Movie}
\item{title}{Title of movie or tv show}
\item{director}{Director name}
\item{cast}{List of cast members}
\item{country}{Countries where movie/tvshow was produced}
\item{date_added}{Date it was added to Netlix}
\item{release_year}{Release year of tv show or movie}
\item{rating}{Maturity rating of of tv show or movie}
\item{duration}{Integer duration of Movie or TV Show (in minutes if movie, in seasons if tv show)}
\item{duration_type}{Duration value of duration variable. Minutes if movie, Seasons if tv show}
\item{listed_in}{List of applicable genres}
\item{description}{Description of movie}
}
}
\source{
The original data set, labeled 'Netflix Movies and TV Shows', can be found on Kaggle
}
\usage{
netflix
}
\description{
This data set consists of TV shows and movies available on Netflix from 2019-January 2021
The data set includes 7787 Movies and TV shows.
}
\examples{
netflix
}
\keyword{datasets}
