% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netflix.R
\name{recommendation_titles}
\alias{recommendation_titles}
\title{Function for getting movie/tv list from a given type, rating and release year}
\usage{
recommendation_titles(type, rating, year)
}
\arguments{
\item{type}{TV show or Movie}

\item{rating}{TV-G, TV-PG, TV-13, TV-14, TV-Y, R, TV-MA}

\item{year}{release year of movie/tv show}
}
\value{
None
}
\description{
Function for getting movie/tv list from a given type, rating and release year
}
\examples{
recommendation_titles("Movie", "TV-MA", 2020)
}
