% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netflix.R
\name{release_year_titles}
\alias{release_year_titles}
\title{Function for getting movie/tv list from a given release year}
\usage{
release_year_titles(year)
}
\arguments{
\item{year}{release year of movies and tv shows on netflix}
}
\value{
None
}
\description{
Function for getting movie/tv list from a given release year
}
\examples{
release_year_titles(2020)
release_year_titles(2010)
}
